@@ -1,17 +1,17 @@
 name: xml2json
-version: 4.4.0
+version: 5.0.0-nullsafety.0
 description: An XML to JSON conversion package. Supports Badgerfish, GData and Parker conversions.
 homepage: https://github.com/shamblett/xml2json
 
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: '>=2.12.0-133.2.beta <3.0.0'
 
 dependencies:
-  xml: '^4.2.0'
+  xml: '^5.0.0-nullsafety.1'
 
 dev_dependencies:
-  test: '^1.15.4'
-  pedantic: '^1.9.1'
+  test: '^1.16.0-nullsafety.13'
+  pedantic: '^1.10.0-nullsafety.3'
   pana: '^0.14.4'
   build_runner: '^1.10.2'
   build_test: '^1.2.0'
