@@ -17,7 +17,7 @@ class _Xml2JsonBadgerfish {
   final String _xmlnsPrefix = '"@xmlns"';
   final String _cdata = '"__cdata"';
 
-  Map<dynamic, dynamic> _transform(XmlDocument node) {
+  Map<dynamic, dynamic> _transform(XmlDocument? node) {
     final json = <dynamic, dynamic>{};
 
     void _process(dynamic node, Map<dynamic, dynamic> obj, dynamic ns) {
@@ -51,7 +51,7 @@ class _Xml2JsonBadgerfish {
           if (name == 'xmlns') {
             ns['$_marker'] = '"$value"';
           } else if (name.indexOf('xmlns:') == 0) {
-            String namePrefix = name.substring(name.indexOf(':') + 1);
+            String? namePrefix = name.substring(name.indexOf(':') + 1);
             namePrefix = '"$namePrefix"';
             ns[namePrefix] = '"$value"';
           } else {
@@ -66,7 +66,7 @@ class _Xml2JsonBadgerfish {
               final pList = <Map<dynamic, dynamic>>[];
               p[_xmlnsPrefix] = pList;
             }
-            final nameMap = <String, String>{};
+            final nameMap = <String, String?>{};
             nameMap[prefix] = ns[prefix];
             p[_xmlnsPrefix].add(nameMap);
           }
@@ -100,7 +100,7 @@ class _Xml2JsonBadgerfish {
   }
 
   /// Transformer function
-  String transform(XmlDocument xmlNode) {
+  String transform(XmlDocument? xmlNode) {
     Map<dynamic, dynamic> json;
     try {
       json = _transform(xmlNode);
